refactor: centralize boolean conversion utility
