+++ backend/controllers/CategoriaController.js
@@ -1,25 +1,11 @@
 const { PrismaClient } = require("@prisma/client");
+const { convertirABoolean } = require("../utils/boolean");
 let prisma;
 try {
   prisma = new PrismaClient();
 } catch (e) {
   prisma = null;
 }
-
-// Utilidad para convertir diferentes entradas a booleano
-const convertirABoolean = (valor) => {
-  if (typeof valor === "boolean") return valor;
-  if (typeof valor === "string") {
-    const lower = valor.toLowerCase();
-    if (lower === "true" || lower === "1" || lower === "activo") return true;
-    if (lower === "false" || lower === "0" || lower === "inactivo") return false;
-  }
-  if (typeof valor === "number") {
-    if (valor === 0) return false;
-    if (valor === 1) return true;
-  }
-  return true;
-};
 
 // Obtener todas las categorías
 const getCategorias = async (req, res) => {
@@ -232,7 +218,6 @@ const toggleEstadoCategoria = async (req, res) => {
 };
 
 module.exports = {
-  convertirABoolean,
   getCategorias,
   getCategoriaById,
   createCategoria,
